requirements.txt
fastapi
uvicorn
torch
transformers
python-dotenv
pydantic
bson
